move DiskGraph struct up one level
