+++ raphtory/src/disk_graph/graph_impl/edge_filter_ops.rs
@@ -1,7 +1,7 @@
-use super::DiskGraph;
 use crate::{
     core::entities::LayerIds,
     db::api::{storage::edges::edge_ref::EdgeStorageRef, view::internal::EdgeFilterOps},
+    disk_graph::DiskGraph,
 };
 
 impl EdgeFilterOps for DiskGraph {
