+++ raphtory/src/lib.rs
@@ -132,7 +132,7 @@ pub use polars_arrow as arrow2;
 #[cfg(test)]
 mod test_utils {
     #[cfg(feature = "storage")]
-    use crate::disk_graph::graph_impl::DiskGraph;
+    use crate::disk_graph::DiskGraph;
     use crate::prelude::Graph;
     #[cfg(feature = "storage")]
     use tempfile::TempDir;
